+++ src/api/user/user.controller.ts
@@ -15,7 +15,7 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import { User } from '@prisma/client';
 import { GetCurrentUser } from 'src/common/decorators/current-user.decorator';
 import { AuthenticateGuard } from 'src/common/guards/authenticate/authenticate.guard';
-import { multerOptions } from 'src/utils/file-upload.utils';
+import { multerOptions } from 'src/common/helper/file-upload';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserService } from './user.service';
